リファクタ: loadCurrentIdToArticleのメソッドは共通の場所に置くようにした
